% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootStrapCurve.R
\name{BootStrapCurve}
\alias{BootStrapCurve}
\title{Title}
\usage{
BootStrapCurve(TargetStruct, InitStructNodes = 15, PlanVarLimit = 0.9,
  PlanVarLimitIC = NULL, nNodes = 30, ForceLasso = FALSE,
  MinBranDiff = 2, nSamples = 10, CellPerc = 0.95, MinProlCells = 10)
}
\arguments{
\item{MinProlCells}{}
}
\description{
Title
}
