% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_quality.R
\name{DistanceGraphCor}
\alias{DistanceGraphCor}
\title{Assess quality of the fitting via correlation}
\usage{
DistanceGraphCor(TargetStruct, ProcStruct, DistVect = 4:10, CorMet = "spe")
}
\arguments{
\item{CorMet}{}
}
\description{
Assess quality of the fitting via correlation
}
