% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_fitting.R
\name{FitCircle}
\alias{FitCircle}
\title{Fit a circle to the data}
\usage{
FitCircle(Data, NonG0Cell, InitStructNodes, PlanVarLimitIC, PlanVarLimit,
  nNodes, Phase1_Overwrite = Inf, Phase2_Overwrite = Inf)
}
\arguments{
\item{PlanVarLimit}{}
}
\description{
Fit a circle to the data
}
