% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_fitting.R
\name{FitLine}
\alias{FitLine}
\title{Fit a line to the data}
\usage{
FitLine(Data, NonG0Cell, InitStructNodes, PlanVarLimitIC, PlanVarLimit, nNodes)
}
\arguments{
\item{PlanVarLimit}{}
}
\description{
Fit a line to the data
}
