% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsePeacks.R
\name{GetGenesWithPeaks}
\alias{GetGenesWithPeaks}
\title{Title}
\usage{
GetGenesWithPeaks(DataStruct, ProcStruct, AllGenes = FALSE, MinNodes = 2,
  FiltMax = 0.2, Thr = 0.9, QuantSel = 0.75, MedianCVThr = 1,
  StagesNames = c("G1", "S", "G2"))
}
\arguments{
\item{SinglePeack}{}
}
\description{
Title
}
