% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootStrapCurve.R
\name{PlotBoot}
\alias{PlotBoot}
\title{Plot bootstrapped data}
\usage{
PlotBoot(SampledStruct, TargetStruct, PoitMult = 2)
}
\arguments{
\item{RefGraphStruct}{}

\item{x}{}
}
\description{
Plot bootstrapped data
}
