% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotOnPseudotime}
\alias{PlotOnPseudotime}
\title{Title}
\usage{
PlotOnPseudotime(WorkStruct, Expression, Name = "", gName, SpanVal = 0.3,
  CatOrder = NULL, legend.position = "bottom")
}
\arguments{
\item{CatOrder}{}
}
\description{
Title
}
