% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotOnStages}
\alias{PlotOnStages}
\title{Title}
\usage{
PlotOnStages(Structure, TaxonList, Categories, PrinGraph, Net, SelThr = NULL,
  ComputeOverlaps = TRUE, RotatioMatrix, PointProjections, ExpData,
  nGenes = 10, OrderOnCat = FALSE, SmoothPoints = 1, PCACenter = FALSE,
  MinCellPerNode = 3, Title = "")
}
\arguments{
\item{ExpData}{}
}
\description{
Title
}
