% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_detection.R
\name{ProjectAndCompute}
\alias{ProjectAndCompute}
\title{Title}
\usage{
ProjectAndCompute(DataSet, GeneSet = NULL, VarThr, nNodes = 30,
  Log = TRUE, Categories = NULL, Filter = TRUE, OutThr = 5,
  PCAFilter = FALSE, OutThrPCA = 5, GraphType = "Circle",
  PlanVarLimit = 0.9, PlanVarLimitIC = NULL, MinBranDiff = 2,
  InitStructNodes = 15, ForceLasso = FALSE, EstProlif = "MeanPerc",
  QuaThr = 0.5, NonG0Cell = NULL, DipPVThr = 0.001, MinProlCells = 20,
  PCACenter = TRUE, PCAProjCenter = FALSE, PlotDebug = FALSE,
  PlotIntermediate = FALSE)
}
\arguments{
\item{PlotIntermediate}{}
}
\description{
Title
}
\examples{

}
