% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{RescaleStagePT}
\alias{RescaleStagePT}
\title{Title}
\usage{
RescaleStagePT(OrderedData, RescaleVal = list(G0 = c(0, 0.25), G1 = c(0.25,
  0.5), S = c(0.5, 0.75), G2 = c(0.75, 1)))
}
\arguments{
\item{RescaleVal}{}
}
\description{
Title
}
