% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneselectiononcurve.R
\name{SelectGenes}
\alias{SelectGenes}
\title{Title}
\usage{
SelectGenes(TaxVect, ExpMat, Mode = "CV", Net = NULL, PriGraph = NULL,
  Proj = NULL)
}
\arguments{
\item{Proj}{}
}
\description{
Title
}
