% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_detection.R
\name{SelectGenesOnGraph}
\alias{SelectGenesOnGraph}
\title{Construct and optimize cycle}
\usage{
SelectGenesOnGraph(DataSet, StartSet, VarThr = 0.99, nNodes = 40,
  Log = TRUE, Categories = NULL, Filter = TRUE, OutThr = 5,
  PCAFilter = TRUE, OutThrPCA = 3, GraphType = "Circle",
  PlanVarLimit = 0.85, PlanVarLimitIC = 0.9, MinBranDiff = 3,
  InitStructNodes = 20, ForceLasso = FALSE, EstProlif = "MeanPerc",
  QuaThr = 0.5, NonG0Cell = NULL, DipPVThr = 1e-04, MinProlCells = 50,
  PCACenter = TRUE, PCAProjCenter = TRUE, PlotDebug = FALSE,
  PlotIntermediate = FALSE, AddGenePerc = 5, SelThr1 = 0.95,
  SelThr2 = 0.99, MadsThr = 1)
}
\arguments{
\item{MadsThr}{}
}
\description{
Construct and optimize cycle
}
