% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsePeacks.R
\name{StageWithPeaks}
\alias{StageWithPeaks}
\title{Title}
\usage{
StageWithPeaks(DataStruct, ProcStruct, StageInfo.UP, StageInfo.DOWN,
  ComputeG0 = TRUE, CCGenes = NULL, MinNodes = 2, FiltMax = 0.2,
  Thr = 0.9, QuantSel = 0.75, MedianCVThr = 1, G0Level = 0.6,
  Mode = 1, Title = "")
}
\arguments{
\item{Mode}{}
}
\description{
Title
}
