% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootStrapCurve.R
\name{BootStrapCurve}
\alias{BootStrapCurve}
\title{Title}
\usage{
BootStrapCurve(TargetStruct, InitStructNodes = 15, PlanVarLimit = 0.85,
  PlanVarLimitIC = 0.9, nNodes = 40, ForceLasso = FALSE,
  MinBranDiff = 2, nSamples = 10, CellPerc = 0.95, MinProlCells = 10,
  VarThr = 1, Cores = 1)
}
\arguments{
\item{MinProlCells}{}
}
\description{
Title
}
