% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootStrapCurve.R
\name{PlotBoot}
\alias{PlotBoot}
\title{Plot bootstrapped data}
\usage{
PlotBoot(SampledStruct, TargetStruct, PoitMult = 1.5, Title = "")
}
\arguments{
\item{TargetStruct}{}
}
\description{
Plot bootstrapped data
}
