% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsePeacks.R
\name{StageWithPeaks}
\alias{StageWithPeaks}
\title{Title}
\usage{
StageWithPeaks(DataStruct, ProcStruct, StageInfo, ComputeG0 = TRUE,
  MinNodes = 2, FiltMax = 0.2, Thr = 0.9, QuantSel = 0.75,
  G0Level = 1, Mode = 1, Title = "")
}
\arguments{
\item{Mode}{}
}
\description{
Title
}
